refactor query min ibc fee helper. improve comments
